feat_#1/_가독성 향상을 위해 AuthDto에서 개별 파일로 분리 및 email 기반으로 수정
