+++ server/src/main/java/oba/backend/server/dto/ReissueRequest.java
@@ -1,0 +1,8 @@
+package oba.backend.server.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record ReissueRequest(
+        @NotBlank String refreshToken
+) {
+}
