+++ server/src/main/java/oba/backend/server/dto/TokenResponse.java
@@ -1,0 +1,8 @@
+package oba.backend.server.dto;
+
+public record TokenResponse(
+        String grantType,
+        String accessToken,
+        String refreshToken
+) {
+}
